some icons. some audio fix.

+++ editor/src/audio_tools/audio_raw_samples.py
@@ -11,11 +11,11 @@ class AudioRawSamples(AudioSegment):
 
     def get_samples_in_between(self, start_time, end_time):
         st = int(start_time*self.sample_rate)
-        et = min((end_time*self.sample_rate)+1, self.samples.shape[1])
+        et = min(int(end_time*self.sample_rate)+1, self.samples.shape[1])
         return self.samples[:, st:et]
 
     def get_samples_in_between_size(self, st, et):
-        return self.samples[:, st:et]
+        return self.samples[:, int(st):int(et)]
 
     def get_index_of_time(self, t):
         index = int(t*self.sample_rate)
